+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
   * New Features:
     * `direct`: Capability to send management messages to `ascp` on running sessions. e.g. change target rate.
+    * `config`: Added command: `sync spec` to get sync parameters documentation. Also added to manual.
   * Issues Fixed:
     * `faspex5`: Fix public link auth for Faspex 5.0.13.
     * `aoc`: Fix some admin operations requiring a user's home for Files.